requirements.txt
Streamlit
langchain
langchain-community
pypdf2
langchain-fireworks
pytesseract
python-docx
pillow
transformers